+++ app/src/main/java/net/opendasharchive/openarchive/FolderAdapter.kt
@@ -1,7 +1,6 @@
 package net.opendasharchive.openarchive
 
 import android.content.Context
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.widget.TextView