+++ app/src/main/java/net/opendasharchive/openarchive/services/gdrive/GDriveActivity.kt
@@ -40,7 +40,7 @@ class GDriveActivity : BaseActivity() {
     private fun removeSpace(space: Space) {
         AlertHelper.show(this, R.string.are_you_sure_you_want_to_remove_this_server_from_the_app, R.string.remove_from_app, buttons = listOf(
             AlertHelper.positiveButton(R.string.remove) { _, _ ->
-                // delete signin from database
+                // delete sign-in from database
                 space.delete()
 
                 // google logout