+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ## 0.3.0 beta8
 - Fixed ANR on Android 14
-- Fixed exessive database reads/writes during file upload
+- Fixed excessive database reads/writes during file upload
 - Added file picker support on more devices
 
 ## 0.3.0 beta7